refactor(CLI): rename lib/cli/cli.js to lib/cli/options.js
The path `lib/cli/cli.js` is confusing. Rename to `options.js` which is
how we were already referencing the file in `lib/cli/etcher.js`.
