+++ lib/cli/etcher.js
@@ -24,7 +24,7 @@ const form = require('resin-cli-form');
 const drivelist = Bluebird.promisifyAll(require('drivelist'));
 const writer = require('./writer');
 const errors = require('./errors');
-const options = require('./cli');
+const options = require('./options');
 const robot = require('../shared/robot');
 const messages = require('../shared/messages');
 const EXIT_CODES = require('../shared/exit-codes');
