+++ lib/cli/options.js
